+++ core/kernel/host_api_builder.cpp
@@ -3,7 +3,6 @@
 
 #include "host_api_builder.hpp"
 
-#include <cstdarg>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>